requirements-test.txt
-r requirements.txt
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-aiohttp>=1.0.0
